CODEOWNERS: splitting files so USM can own its own files
