+++ pkg/network/encoding/format_test.go
@@ -6,7 +6,6 @@
 package encoding
 
 import (
-	"github.com/stretchr/testify/assert"
 	"runtime"
 	"testing"
 
@@ -90,45 +89,4 @@ func BenchmarkConnectionReset(b *testing.B) {
 		c.Reset()
 	}
 	runtime.KeepAlive(c)
-}
-
-func TestFormatProtocols(t *testing.T) {
-	tests := []struct {
-		name     string
-		protocol network.ProtocolType
-		want     *model.ProtocolStack
-	}{
-		{
-			name:     "unknown protocol",
-			protocol: network.ProtocolUnknown,
-			want: &model.ProtocolStack{
-				Stack: []model.ProtocolType{
-					model.ProtocolType_protocolUnknown,
-				},
-			},
-		},
-		{
-			name:     "unclassified protocol",
-			protocol: network.ProtocolUnclassified,
-			want: &model.ProtocolStack{
-				Stack: []model.ProtocolType{
-					model.ProtocolType_protocolUnknown,
-				},
-			},
-		},
-		{
-			name:     "http protocol",
-			protocol: network.ProtocolHTTP,
-			want: &model.ProtocolStack{
-				Stack: []model.ProtocolType{
-					model.ProtocolType_protocolHTTP,
-				},
-			},
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			assert.Equalf(t, tt.want, formatProtocol(tt.protocol), "formatProtocol(%v)", tt.protocol)
-		})
-	}
 }
