+++ pkg/network/event_common.go
@@ -18,17 +18,6 @@ import (
 	"github.com/DataDog/datadog-agent/pkg/process/util"
 )
 
-type ProtocolType uint16
-
-const (
-	ProtocolUnclassified ProtocolType = iota
-	ProtocolUnknown
-	ProtocolHTTP
-	ProtocolHTTP2
-	ProtocolTLS
-	MaxProtocols
-)
-
 // ConnectionType will be either TCP or UDP
 type ConnectionType uint8
